update createSensor method

+++ src/main/java/com/loschiferos/ztech/pot/domain/model/commands/CreateSensorCommand.java
@@ -1,4 +1,4 @@
 package com.loschiferos.ztech.pot.domain.model.commands;
 
-public record CreateSensorCommand(Long flowerpotId, Long type, Double value) {
+public record CreateSensorCommand(Long flowerpotId, Long type, Long value) {
 }
