+++ src/main/java/com/loschiferos/ztech/pot/interfaces/rest/resources/CreateSensorResource.java
@@ -1,4 +1,4 @@
 package com.loschiferos.ztech.pot.interfaces.rest.resources;
 
-public record CreateSensorResource(Long flowerpotId, Long type, Double value) {
+public record CreateSensorResource(Long flowerpotId, Long type, Long value) {
 }
